Look save into one fader! beta
Changed how look saving/loading works.  A look is now saved as an
extension of Beam - a Look.  Since a Look is a kind of Beam, we can
throw them around easily.  The ability to break a look back out into
constituent components is not implemented yet.  Nor is this build
debugged yet, consider it beta.  Also fixed a few default behaviors.